code clean

+++ Share/Models/Common/KeyValueDescription.cs
@@ -2,7 +2,7 @@ namespace Share.Models.Common
 {
     public class KeyValueDescription
     {
-        public int Key { get; set; }
+        public byte Key { get; set; }
         public string Value { get; set; }
         public string Description { get; set; }
     }
